requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.12
h11==0.16.0
idna==3.10
MarkupSafe==3.0.2
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.4.8
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.1.0
python-http-client==3.3.7
python-jose==3.4.0
python-multipart==0.0.20
rsa==4.9.1
sendgrid==6.12.3
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
Werkzeug==3.1.3
